test(bb-knowledge-base): consolidate KB fix tests into canonical test files
