+++ packages/bb-knowledge-base/src/index.mock.test.ts
@@ -6,6 +6,7 @@ import assert from 'node:assert';
 import { rmSync, existsSync, mkdirSync, writeFileSync, cpSync } from 'node:fs';
 import { join, dirname } from 'node:path';
 import { fileURLToPath } from 'node:url';
+import { spawnSync } from 'node:child_process';
 import { KnowledgeBase, KnowledgeBaseErrors } from './index.mock.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
@@ -22,6 +23,23 @@ function setupTestFixtures(): void {
 function cleanup(): void {
 	try { rmSync('.bb-data', { recursive: true, force: true }); } catch {}
 	try { rmSync(TEST_KNOWLEDGE, { recursive: true, force: true }); } catch {}
+}
+
+// Helpers for tests that build their own temporary corpora under the project
+// directory (path-guard, cache, filter, load-recovery, chunking, unicode).
+const CWD = process.cwd();
+
+function freshDir(name: string): string {
+	const p = join(CWD, name);
+	if (existsSync(p)) rmSync(p, { recursive: true, force: true });
+	mkdirSync(p, { recursive: true });
+	return p;
+}
+
+function cleanupDirs(...names: string[]): void {
+	for (const name of names) {
+		try { rmSync(join(CWD, name), { recursive: true, force: true }); } catch {}
+	}
 }
 
 beforeEach(() => {
@@ -430,6 +448,324 @@ test('metadata.json sidecar files are not indexed as documents', async () => {
 			`should not index .metadata.json files as documents, but found: ${r.source}`,
 		);
 	}
+});
+
+// ── Source path containment (path-guard) ───────────────────────────────────
+//
+// The mock must reject sources that escape the project directory. The guard
+// uses path-separator-aware comparison so a sibling directory that merely
+// shares the cwd string prefix (e.g. /a/proj-secrets vs /a/proj) is NOT
+// treated as inside the project.
+
+describe('source path containment', () => {
+	test('rejects sibling directory that shares the cwd string prefix', () => {
+		const base = freshDir('_path-guard-base');
+		const proj = join(base, 'proj');
+		mkdirSync(proj, { recursive: true });
+		const sibling = join(base, 'proj-secrets');
+		mkdirSync(sibling, { recursive: true });
+		writeFileSync(
+			join(sibling, 'secret.md'),
+			'This is a secret document that should never be accessible from the project directory.',
+		);
+
+		try {
+			// Run in a subprocess with cwd=proj so resolve(cwd) = .../proj. __dirname
+			// resolves to the compiled dist/ directory, where index.mock.js sits beside
+			// this test, so the import path stays valid after consolidation.
+			const distPath = join(__dirname, 'index.mock.js').replace(/\\/g, '/');
+			const script = `
+import { KnowledgeBase, KnowledgeBaseErrors } from '${distPath}';
+try {
+	const kb = new KnowledgeBase({ id: 'app' }, 'leak', { source: '../proj-secrets' });
+	await kb.retrieve('secret document');
+	process.exit(1); // retrieve resolved — guard failed to block
+} catch (e) {
+	if (e.name === '${KnowledgeBaseErrors.InvalidSource}') {
+		process.exit(0); // correctly blocked
+	}
+	process.exit(2); // unexpected error type
+}
+`;
+			const res = spawnSync(process.execPath, ['--input-type=module', '-e', script], {
+				cwd: proj,
+				encoding: 'utf8',
+				timeout: 10000,
+			});
+
+			assert.strictEqual(
+				res.status,
+				0,
+				`Expected InvalidSource for sibling dir. Exit code ${res.status}. stdout: ${res.stdout}. stderr: ${res.stderr}`,
+			);
+		} finally {
+			rmSync(base, { recursive: true, force: true });
+		}
+	});
+
+	test('rejects source with .. that escapes the project directory', async () => {
+		const outside = freshDir('_path-guard-outside');
+		writeFileSync(
+			join(outside, 'data.txt'),
+			'Sensitive data that lives outside the project root and must not be reachable via relative paths.',
+		);
+
+		try {
+			const kb = new KnowledgeBase({ id: 'app' }, 'escape', {
+				source: join('..', '_path-guard-outside'),
+			});
+			await assert.rejects(
+				() => kb.retrieve('sensitive data'),
+				(err: Error) => {
+					assert.strictEqual(err.name, KnowledgeBaseErrors.InvalidSource);
+					return true;
+				},
+				'Should throw InvalidSource for paths escaping project via ..',
+			);
+		} finally {
+			rmSync(outside, { recursive: true, force: true });
+		}
+	});
+
+	test('rejects absolute path outside the project', async () => {
+		const outside = freshDir('_path-guard-abs');
+		writeFileSync(join(outside, 'doc.md'), 'Absolute path document content for testing path guard enforcement.');
+
+		try {
+			const kb = new KnowledgeBase({ id: 'app' }, 'abs', { source: outside });
+			await assert.rejects(
+				() => kb.retrieve('absolute path document'),
+				(err: Error) => {
+					assert.strictEqual(err.name, KnowledgeBaseErrors.InvalidSource);
+					return true;
+				},
+				'Should throw InvalidSource for absolute paths outside project',
+			);
+		} finally {
+			rmSync(outside, { recursive: true, force: true });
+		}
+	});
+});
+
+// ── Cache invalidation ──────────────────────────────────────────────────────
+//
+// The chunks cache is keyed on the source folder's contents (path plus per-file
+// mtime/size). Changing the source must rebuild the index rather than serving
+// stale results.
+
+describe('cache invalidation', () => {
+	test('switching the source folder returns new content, not stale cache', async () => {
+		const v1 = freshDir('_cache-v1');
+		writeFileSync(join(v1, 'doc.md'), 'Version one content about apples oranges and fruit orchards in the countryside.');
+		const v2 = freshDir('_cache-v2');
+		writeFileSync(join(v2, 'doc.md'), 'Version two content about spaceships rockets and interstellar galactic travel.');
+
+		try {
+			const kb1 = new KnowledgeBase({ id: 'app' }, 'cache', { source: '_cache-v1' });
+			const r1 = await kb1.retrieve('apples oranges fruit');
+			assert.ok(r1.length > 0, 'v1 query should return results');
+			assert.ok(
+				r1[0].text.includes('apples') || r1[0].text.includes('oranges'),
+				'v1 results should contain v1 content',
+			);
+
+			// Same fullId, different source folder — must not reuse the v1 cache.
+			const kb2 = new KnowledgeBase({ id: 'app' }, 'cache', { source: '_cache-v2' });
+			const r2 = await kb2.retrieve('spaceships rockets galactic');
+			assert.ok(r2.length > 0, 'v2 query should return results when source points to v2');
+			assert.ok(
+				r2[0].text.includes('spaceships') || r2[0].text.includes('rockets'),
+				'results should contain v2 content, not stale v1 data',
+			);
+		} finally {
+			cleanupDirs('_cache-v1', '_cache-v2');
+		}
+	});
+
+	test('editing a document in the source folder is reflected on next load', async () => {
+		const src = freshDir('_cache-edit');
+		writeFileSync(join(src, 'doc.md'), 'Original content about dinosaurs and prehistoric creatures from the Jurassic period.');
+
+		try {
+			const kb1 = new KnowledgeBase({ id: 'app' }, 'edit', { source: '_cache-edit' });
+			const r1 = await kb1.retrieve('dinosaurs prehistoric Jurassic');
+			assert.ok(r1.length > 0, 'original query should match');
+
+			writeFileSync(join(src, 'doc.md'), 'Updated content about quantum computing and artificial intelligence breakthroughs in modern science.');
+
+			const kb2 = new KnowledgeBase({ id: 'app' }, 'edit', { source: '_cache-edit' });
+			const r2 = await kb2.retrieve('quantum computing artificial intelligence');
+			assert.ok(r2.length > 0, 'updated content should return results after edit');
+			assert.ok(
+				r2[0].text.includes('quantum') || r2[0].text.includes('artificial'),
+				'results should reflect the edited content',
+			);
+
+			const r3 = await kb2.retrieve('dinosaurs prehistoric Jurassic');
+			assert.strictEqual(r3.length, 0, 'old content should not be returned after the document edit');
+		} finally {
+			cleanupDirs('_cache-edit');
+		}
+	});
+});
+
+// ── Metadata filter completeness ────────────────────────────────────────────
+//
+// When a filter is supplied the mock scores every chunk before filtering, so a
+// matching doc is never silently dropped just because higher-scoring
+// non-matching docs would have filled the maxResults window first.
+
+describe('metadata filter completeness', () => {
+	test('filter returns all matching docs even when they rank below higher-scoring non-matching docs', async () => {
+		const src = freshDir('_filter-completeness');
+		mkdirSync(join(src, 'target'), { recursive: true });
+		mkdirSync(join(src, 'noise'), { recursive: true });
+
+		// 200 noise docs that strongly match "apple" (high TF-IDF score).
+		for (let i = 0; i < 200; i++) {
+			writeFileSync(
+				join(src, 'noise', `n${i}.md`),
+				'apple apple apple apple apple fruit orchard harvest season picking baskets cider juice.',
+			);
+		}
+		// 5 target docs in folder=target that mention "apple" with lower density.
+		for (let i = 0; i < 5; i++) {
+			writeFileSync(
+				join(src, 'target', `t${i}.md`),
+				`Target document ${i} that mentions apple once amid many other unrelated words and topics and themes and ideas and concepts here.`,
+			);
+		}
+
+		try {
+			const kb = new KnowledgeBase({ id: 'app' }, 'filtcomp', { source: '_filter-completeness' });
+			const results = await kb.retrieve('apple', {
+				maxResults: 5,
+				filter: { folder: { equals: 'target' } },
+			});
+
+			assert.strictEqual(
+				results.length,
+				5,
+				`Expected 5 results from target folder, got ${results.length}. ` +
+					'Filter must scan all matching docs regardless of score ranking.',
+			);
+			for (const r of results) {
+				assert.strictEqual(r.metadata.folder, 'target');
+			}
+		} finally {
+			cleanupDirs('_filter-completeness');
+		}
+	});
+});
+
+// ── Load recovery after a failed load ────────────────────────────────────────
+//
+// A failed load must not permanently poison the instance: the internal
+// loadPromise is reset on failure so a later retrieve() can succeed once the
+// underlying problem (e.g. a missing source folder) is resolved.
+
+describe('load recovery after failed load', () => {
+	test('instance recovers after the source folder is created following an initial failure', async () => {
+		const srcName = '_error-recovery-src';
+		cleanupDirs(srcName);
+
+		const kb = new KnowledgeBase({ id: 'app' }, 'recover', { source: srcName });
+		await assert.rejects(
+			() => kb.retrieve('test query'),
+			(err: Error) => {
+				assert.strictEqual(err.name, KnowledgeBaseErrors.InvalidSource);
+				return true;
+			},
+			'First call should fail because the source folder does not exist',
+		);
+
+		const src = freshDir(srcName);
+		writeFileSync(
+			join(src, 'doc.md'),
+			'This document contains information about recovery testing and transient failures in distributed systems.',
+		);
+
+		try {
+			const results = await kb.retrieve('recovery testing transient failures');
+			assert.ok(
+				results.length > 0,
+				'Instance should recover after the source folder is created — loadPromise must not be permanently poisoned',
+			);
+		} finally {
+			cleanupDirs(srcName);
+		}
+	});
+});
+
+// ── Chunking configuration ──────────────────────────────────────────────────
+//
+// The 'fixed' chunking strategy must be honored: a small chunkSize produces
+// smaller chunks than the default 'semantic' strategy, which returns a
+// blank-line-free document as a single oversized chunk.
+
+describe('chunking configuration', () => {
+	test('fixed chunking with a small chunkSize produces smaller chunks than the default', async () => {
+		const src = freshDir('_chunking-fixed');
+		// Single large paragraph (no blank lines) — about 500 words.
+		const words = Array.from(
+			{ length: 500 },
+			(_, i) => `word${i} sentence${i % 50} paragraph content text document knowledge base`,
+		).join(' ');
+		writeFileSync(join(src, 'big.md'), words);
+
+		try {
+			const kbDefault = new KnowledgeBase({ id: 'app' }, 'chunkdef', { source: '_chunking-fixed' });
+			const defaultResults = await kbDefault.retrieve('word0 sentence0');
+
+			const kbFixed = new KnowledgeBase({ id: 'app' }, 'chunkfix', {
+				source: '_chunking-fixed',
+				chunking: { strategy: 'fixed', chunkSize: 50 },
+			});
+			const fixedResults = await kbFixed.retrieve('word0 sentence0');
+
+			if (defaultResults.length > 0 && fixedResults.length > 0) {
+				const defaultMaxLen = Math.max(...defaultResults.map((r) => r.text.length));
+				const fixedMaxLen = Math.max(...fixedResults.map((r) => r.text.length));
+				assert.ok(
+					fixedMaxLen < defaultMaxLen,
+					`Fixed chunking (chunkSize=50) should produce smaller chunks than default. ` +
+						`Fixed max: ${fixedMaxLen}, Default max: ${defaultMaxLen}. ` +
+						'The chunking option appears to be ignored.',
+				);
+			} else {
+				assert.ok(fixedResults.length > 0, 'Fixed chunking should still produce searchable results');
+			}
+		} finally {
+			cleanupDirs('_chunking-fixed');
+		}
+	});
+});
+
+// ── Unicode / multilingual retrieval ─────────────────────────────────────────
+//
+// The TF-IDF tokenizer is Unicode-aware (accent folding plus CJK bigrams). This
+// end-to-end test proves non-ASCII content survives the file → chunk → index →
+// retrieve pipeline. Accent-folding specifics are unit-tested in tfidf.test.ts.
+
+describe('unicode / multilingual retrieval', () => {
+	test('CJK content is searchable end-to-end', async () => {
+		const src = freshDir('_unicode-cjk');
+		writeFileSync(
+			join(src, 'doc.md'),
+			'機械学習のアルゴリズムはデータ処理を効率化します。人工知能は未来の技術です。',
+		);
+
+		try {
+			const kb = new KnowledgeBase({ id: 'app' }, 'uni-cjk', { source: '_unicode-cjk' });
+			const results = await kb.retrieve('機械学習 人工知能');
+			assert.ok(
+				results.length > 0,
+				'CJK query should find CJK document content — tokenizer must not strip non-ASCII characters entirely',
+			);
+		} finally {
+			cleanupDirs('_unicode-cjk');
+		}
+	});
 });
 
 // ── Cleanup after all tests ────────────────────────────────────────────────
